chore(testing): move story utils into common util area
