+++ packages/angular/src/generators/component-test/component-test.ts
@@ -1,21 +1,16 @@
-import { getModuleDeclaredComponents } from '@nrwl/angular/src/generators/stories/lib/module-info';
-import type { PropertyDeclaration } from 'typescript';
-import { getComponentProps } from '../utils/storybook';
 import { assertMinimumCypressVersion } from '@nrwl/cypress/src/utils/cypress-version';
+import { generateFiles, joinPathFragments, Tree } from '@nrwl/devkit';
 import {
-  generateFiles,
-  joinPathFragments,
-  readProjectConfiguration,
-  Tree,
-} from '@nrwl/devkit';
-import { dirname, relative, basename, extname } from 'path';
+  getArgsDefaultValue,
+  getComponentProps,
+} from '../utils/storybook-ast/storybook-inputs';
 import { ComponentTestSchema } from './schema';
+
 export function componentTestGenerator(
   tree: Tree,
   options: ComponentTestSchema
 ) {
-  // TODO: enable this assert
-  // assertMinimumCypressVersion(10);
+  assertMinimumCypressVersion(10);
   const componentDirPath = joinPathFragments(
     options.projectPath,
     options.componentDir
@@ -49,18 +44,4 @@ export function componentTestGenerator(
       }
     );
   }
-}
-
-// TODO: pulled this from component-story
-function getArgsDefaultValue(property: PropertyDeclaration): string {
-  const typeNameToDefault = {
-    string: "''",
-    number: '0',
-    boolean: 'false',
-  };
-  return property.initializer
-    ? property.initializer.getText()
-    : property.type
-    ? typeNameToDefault[property.type.getText()]
-    : "''";
 }
